Changed Layout + Sidebar + Fix SearchBar
