+++ package.json
@@ -34,7 +34,7 @@
     "@fortawesome/react-fontawesome": "^0.1.9",
     "@material-ui/core": "^4.9.14",
     "@material-ui/lab": "^4.0.0-alpha.56",
-    "@terminal-packages/space-ui": "^0.3.10",
+    "@terminal-packages/space-ui": "^0.3.11",
     "@toruslabs/torus-direct-web-sdk": "^3.4.1",
     "axios": "^0.20.0",
     "chalk": "^4.1.0",
