--- src/shared/components/Layout/components/Account/constants.js
@@ -1,6 +1,0 @@
-// eslint-disable-next-line import/prefer-default-export
-export const MENU_DROPDOWN_ITEMS = {
-  help: 'help',
-  signout: 'signout',
-  settings: 'settings',
-};
