--- src/shared/components/Layout/components/CreateNewButton/constants.js
@@ -1,6 +1,0 @@
-// eslint-disable-next-line import/prefer-default-export
-export const MENU_DROPDOWN_ITEMS = {
-  newFolder: 'new-folder',
-  fileUpload: 'file-upload',
-  folderUpload: 'folder-upload',
-};
