+++ src/views/Storage/shared/components/HeaderNav/index.js
@@ -78,12 +78,6 @@ const HeaderNav = () => {
     </Typography>
   );
 
-  const searchResults = results ? results.map((item) => ({
-    ext: item.ext,
-    key: item.id,
-    fileName: item.key,
-  })) : null;
-
   return (
     <div className={classes.root}>
       <FolderNavButton
@@ -105,12 +99,19 @@ const HeaderNav = () => {
       />
       <SearchBar
         debounce={200}
-        results={searchResults}
+        results={results}
         value={searchTerm}
         noResults={noResults}
         placeholder={t('common.search')}
         onChange={(val) => searchFiles(val)}
+        mapItemToFileItemProps={(item) => ({
+          ext: item.ext,
+          key: item.id,
+          fileName: item.key,
+        })}
         onClickResult={(item, hideResults) => {
+          /* eslint-disable */
+          console.log('item', item);
           navigateToFolder(item, history);
           openItem(item);
           hideResults();
